Use safe path for project owner name
`project.namespace['name']` can be values such as 'A Name' or 'Namé',
which are modified by name_clean(). However, if the pre-safed path
from GitLab is use from the outset, fewer inconsistencies occur.

Eg., A project by "Orion F" with a namespace path "orionf" would cause
collaborators to be searched for on the Gitea side using "Orion_F/...".